Fixed: test file for npc
