+++ src/player.py
@@ -3,9 +3,133 @@ import random
 import bisect
 import rank
 import numpy as np
+from dataclasses import dataclass
 from deck import Deck, Card
-from poker_game import Player
+from rank import BestHand
 
+
+@dataclass
+class Player:
+    name: str = random.sample(["Henry", "Jiachen", "Jonathan"], 1)[0]
+    stack: int = 0
+    hand = []
+    bet_this_round: int = 0
+    current_decision: int | str | None = None
+    status: (
+        str | None
+    ) = None  # 'little blind', 'big blind', 'dealer' 'highest bettor'
+
+    def __repr__(self) -> str:
+        return str((self.name, self.stack, self.hand))
+
+    def draw_hand(self, deck: Deck) -> None:
+        self.hand = [deck.draw() for _ in range(2)]
+
+    def clear_hand(self) -> None:
+        self.hand = []
+
+    def best_hand(self, board: list[Card]) -> BestHand:
+        return BestHand(self.hand, board)
+
+    def reset_action(self):
+        self.status = None
+        self.current_decision = None
+        self.bet_this_round = 0
+
+    def bet(self, action: int) -> None:
+        """
+        Sets current_decision and bet_this_round attributs on self
+        """
+        self.current_decision = action
+        self.bet_this_round += self.current_decision
+
+    def decision(
+        self,
+        table_cards: list,
+        current_bet: int,
+        pot: int,
+        asking_again_message: str = None,
+    ) -> None:
+        """
+        Decision method prompts the user to choose an action for their turn.
+        """
+        # Skip player who is all-in
+        if self.stack == 0:
+            self.current_decision = 0
+            return
+
+        if asking_again_message is None:
+            action = input(
+                f"Awaiting {self.name}'s decision...\n"
+                "\tcards on table:\t" + " ".join(map(str, table_cards)) + "\n"
+                f"\tpot:\t\t${pot}\n"
+                f"\tcurrent bet:\t${current_bet}\n"
+                "\n"
+                "\thand:\t\t" + " ".join(map(str, self.hand)) + "\n"
+                f"\tstack:\t\t${self.stack}\n"
+                f"\talready bet:\t${self.bet_this_round}\n"
+                "Please make a decision: "
+            )
+        else:
+            action = input(asking_again_message)
+        try:
+            action = int(action)
+            if current_bet <= action <= self.stack:
+                self.bet(action - self.bet_this_round)
+            elif action < current_bet:
+                self.decision(
+                    table_cards,
+                    current_bet,
+                    pot,
+                    asking_again_message="Your bet ain't high enough, cowbo"
+                    "y. Try again: ",
+                )
+            elif action - self.bet_this_round > self.stack:
+                self.decision(
+                    table_cards,
+                    current_bet,
+                    pot,
+                    asking_again_message="Not enough chips in the stack for"
+                    " that one. Please bet a valid amount: ",
+                )
+        except ValueError:
+            if action.upper() == "CALL":
+                if self.bet_this_round < current_bet <= self.stack:
+                    self.bet(current_bet - self.bet_this_round)
+                elif current_bet == 0:
+                    self.decision(
+                        table_cards,
+                        current_bet,
+                        pot,
+                        asking_again_message="You cannot call. Try something "
+                        "else: ",
+                    )
+                else:
+                    self.bet(self.stack)
+            elif action.upper() == "CHECK":
+                if current_bet == self.bet_this_round:
+                    self.current_decision = 0
+                    self.status = None
+                else:
+                    self.decision(
+                        table_cards,
+                        current_bet,
+                        pot,
+                        asking_again_message="You cannot check. Try something"
+                        " else: ",
+                    )
+            elif action.upper() == "ALL IN":
+                self.bet(self.stack)
+            elif action.upper() == "FOLD":
+                self.current_decision = "FOLD"
+            else:
+                self.decision(
+                    table_cards,
+                    current_bet,
+                    pot,
+                    asking_again_message="Invalid decision. Try something els"
+                    "e: ",
+                )
 
 class NpcRandom(Player):
     """
